requirements.txt
flask
pdfplumber
python-docx
scikit-learn
